Added use of ffprobe
Utilize PathBuf manipulation to identify if there is a compressed
version of an mkv file.
When the suspected compressed sibling file exists, use ffprobe to
verify that it is encoded with the desired HEVC codec and has the
hvc1 codec tag.
Using the combination of this data to decide whether or not to
flag the file as compressed.