Switch to PollWatcher
The recommended watcher was only producing events from the host
for a CIFS mount; changed to PollWatcher.

Bumped transcode parameters to slow down the transcoder since we only
get notified when a file is created, not when it is closed by a writer.

Handle SIGINT during the initial directory scan.
Gracefully handle the situation when an enqueued file is removed prior
to dequeue.